requirements.txt
apache-airflow==2.9.2
requests==2.32.3
python-dotenv==1.0.1
openai==1.35.3
anthropic==0.34.1
unstructured==0.15.1
unstructured-inference==0.7.36
pypdf==4.3.1
langchain_community==0.2.6
langchain_milvus==0.1.4
PyLD==2.0.4
jiter==0.5.0
langchain-core==0.2.36
pymilvus==2.4.6
pymysql==1.0.3
SQLAlchemy>=1.4.36,<2.0
flask_cors==5.0.0
kubernetes==31.0.0
apache-airflow-providers-mysql==5.7.1
mysqlclient==2.2.4
